Struct jwalk::WalkDir

pub struct WalkDir { /* fields omitted */ }

Builder for walking a directory.

Implementations

impl WalkDir

pub fn new<P: AsRef<Path>>(root: P) -> Self

Create a builder for a recursive directory iterator starting at the file path root. If root is a directory, then it is the first item yielded by the iterator. If root is a file, then it is the first and only item yielded by the iterator.

pub fn root(&self) -> &Path

Root path of the walk.

pub fn sort(self, sort: bool) -> Self

Sort entries by file_name per directory. Defaults to false. Use process_entries for custom sorting or filtering.

pub fn skip_hidden(self, skip_hidden: bool) -> Self

Skip hidden entries. Enabled by default.

pub fn preload_metadata(self, preload_metadata: bool) -> Self

Preload metadata before yielding entries. When running in parrallel the metadata is loaded in rayon's thread pool.

pub fn max_depth(self, depth: usize) -> Self

Maximum depth of entries yielded by the iterator. 0 corresponds to the root path of this walk.

A depth < 2 will automatically change thread_count to 1. Parrallelism happens at the fs::read_dir level. It only makes sense to use multiple threads when reading more then one directory.

pub fn num_threads(self, n: usize) -> Self

0 Use rayon global pool.

1 Perform walk on calling thread.

n > 1 Construct a new rayon ThreadPool to perform the walk.

pub fn process_entries<F>(self, process_by: F) -> Self where
F: Fn(&mut Vec<Result<DirEntry>>) + Send + Sync + 'static,

A callback function to process (sort/filter/skip) each directory of entries before they are yielded. Modify the given array to sort/filter entries. Use entry.content_spec = None to yield an entry but skip reading its contents.